+++ interactive.py
@@ -13,27 +13,27 @@ from spritesheet_functions import SpriteSheet
 # at least in 320 I guess it's 32x32 in 640
 """globals: walls, exit_blocks"""
 
-class GameView(object):
-    """this view handles displaying most of the things.
+# class GameView(object):
+#     """this view handles displaying most of the things.
 
-    attributes: model, screen"""
-    def __init__(self, model, size):
-        self.model = model
-        self.screen = pygame.display.set_mode(size)
-        pygame.mixer.music.load('sounds/TheDarkLake.mp3') #this music plays
-        pygame.mixer.music.play(-1)
+#     attributes: model, screen"""
+#     def __init__(self, model, size):
+#         self.model = model
+#         self.screen = pygame.display.set_mode(size)
+#         pygame.mixer.music.load('sounds/TheDarkLake.mp3') #this music plays
+#         pygame.mixer.music.play(-1)
 
 
-    def draw(self):
-        """ Draw the game to the pygame window"""
-        background = pygame.image.load('images/room2.png').convert()
-        scaled_bg = pygame.transform.scale(background, size)
-        self.screen.blit(scaled_bg, (0,0))
-        char = self.model.char.image
-        self.screen.blit(char, (self.model.char.x,self.model.char.y))
-        pygame.draw.rect(self.screen, (255, 200, 0), self.model.char.rect)
-        # draws a yellow box where the rect is
-        pygame.display.flip()
+#     def draw(self):
+#         """ Draw the game to the pygame window"""
+#         background = pygame.image.load('images/room2.png').convert()
+#         scaled_bg = pygame.transform.scale(background, size)
+#         self.screen.blit(scaled_bg, (0,0))
+#         char = self.model.char.image
+#         self.screen.blit(char, (self.model.char.x,self.model.char.y))
+#         pygame.draw.rect(self.screen, (255, 200, 0), self.model.char.rect)
+#         # draws a yellow box where the rect is
+#         pygame.display.flip()
 
 
 class GameModel(object):
@@ -158,6 +158,8 @@ class CharacterSprite(pygame.sprite.Sprite):
 
     """These functions move the thing. Should check for collisions and adjust
     accordingly now? Untested cause the wall rects don't exist yet"""
+
+    #These should really be one move function with two parameters...
     def move_left(self, li):
         self.x -= self.step_size
         self.rect.move_ip(-self.step_size, 0)
